rework tty readstream and writestream

+++ ext/node/polyfills/internal_binding/stream_wrap.ts
@@ -91,7 +91,7 @@ export const kLastWriteWasAsync = StreamBaseStateFields.kLastWriteWasAsync;
 export const kNumStreamBaseStateFields =
   StreamBaseStateFields.kNumStreamBaseStateFields;
 
-export const streamBaseState = new Uint8Array(5);
+export const streamBaseState = new Int32Array(5);
 
 // This is Deno, it always will be async.
 streamBaseState[kLastWriteWasAsync] = 1;
